+++ packages/article-magazine-standard/__tests__/mocks.web.js
@@ -3,6 +3,7 @@ jest.mock("@times-components/ad", () => require("./ad-mock"));
 jest.mock("@times-components/article-byline", () => ({
   ArticleBylineWithLinks: "ArticleBylineWithLinks"
 }));
+jest.mock("@times-components/article-comments", () => "ArticleComments");
 jest.mock("@times-components/article-flag", () => ({
   ExclusiveArticleFlag: "ExclusiveArticleFlag",
   NewArticleFlag: "NewArticleFlag",