+++ packages/article-main-standard/__tests__/web/semantic.web.test.js
@@ -143,6 +143,7 @@ const tests = [
             onRelatedArticlePress={() => {}}
             onTopicPress={() => {}}
             onVideoPress={() => {}}
+            spotAccountId=""
           />
         </Context.Provider>
       );