+++ packages/ssr/src/client/article.js
@@ -2,14 +2,20 @@ const article = require("../component/article");
 const runClient = require("../lib/run-client");
 
 if (window.nuk && window.nuk.ssr && window.nuk.article) {
-  const { rootTag, makeArticleUrl, mapArticleToAdConfig } = window.nuk.ssr;
+  const {
+    rootTag,
+    makeArticleUrl,
+    mapArticleToAdConfig,
+    spotAccountId
+  } = window.nuk.ssr;
   const { articleId, debounceTimeMs } = window.nuk.article;
 
   const data = {
     articleId,
     debounceTimeMs,
     makeArticleUrl,
-    mapArticleToAdConfig
+    mapArticleToAdConfig,
+    spotAccountId
   };
 
   const clientOptions = {