+++ packages/article-magazine-comment/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@0.1.23...@times-components/article-magazine-comment@0.2.0) (2018-12-12)
+
+
+### Features
+
+* add SpotIM comments to articles on web ([#1562](https://github.com/newsuk/times-components/issues/1562)) ([52dc2d7](https://github.com/newsuk/times-components/commit/52dc2d7))
+
+
+
+
+
+## [0.1.23](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@0.1.22...@times-components/article-magazine-comment@0.1.23) (2018-12-12)
+
+
+### Bug Fixes
+
+* magazine comment empty author ([#1568](https://github.com/newsuk/times-components/issues/1568)) ([37545b4](https://github.com/newsuk/times-components/commit/37545b4))
+
+
+
+
+
+## [0.1.22](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@0.1.21...@times-components/article-magazine-comment@0.1.22) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [0.1.21](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@0.1.20...@times-components/article-magazine-comment@0.1.21) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [0.1.20](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@0.1.19...@times-components/article-magazine-comment@0.1.20) (2018-12-11)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
 ## [0.1.19](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@0.1.18...@times-components/article-magazine-comment@0.1.19) (2018-12-11)
 
 **Note:** Version bump only for package @times-components/article-magazine-comment