+++ packages/article-main-comment/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@0.3.21...@times-components/article-main-comment@0.4.0) (2018-12-12)
+
+
+### Features
+
+* add SpotIM comments to articles on web ([#1562](https://github.com/newsuk/times-components/issues/1562)) ([52dc2d7](https://github.com/newsuk/times-components/commit/52dc2d7))
+
+
+
+
+
+## [0.3.21](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@0.3.20...@times-components/article-main-comment@0.3.21) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [0.3.20](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@0.3.19...@times-components/article-main-comment@0.3.20) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [0.3.19](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@0.3.18...@times-components/article-main-comment@0.3.19) (2018-12-11)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
 ## [0.3.18](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@0.3.17...@times-components/article-main-comment@0.3.18) (2018-12-11)
 
 **Note:** Version bump only for package @times-components/article-main-comment