+++ packages/brightcove-video/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.63](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@3.1.62...@times-components/brightcove-video@3.1.63) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
+
+
+
+
+## [3.1.62](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@3.1.61...@times-components/brightcove-video@3.1.62) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
+
+
+
+
 ## [3.1.61](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@3.1.60...@times-components/brightcove-video@3.1.61) (2018-12-06)
 
 **Note:** Version bump only for package @times-components/brightcove-video