+++ packages/card/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.27](https://github.com/newsuk/times-components/compare/@times-components/card@4.2.26...@times-components/card@4.2.27) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/card
+
+
+
+
+
+## [4.2.26](https://github.com/newsuk/times-components/compare/@times-components/card@4.2.25...@times-components/card@4.2.26) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/card
+
+
+
+
+
 ## [4.2.25](https://github.com/newsuk/times-components/compare/@times-components/card@4.2.24...@times-components/card@4.2.25) (2018-12-06)
 
 **Note:** Version bump only for package @times-components/card