+++ packages/depend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.13](https://github.com/newsuk/times-components/compare/@times-components/depend@2.1.12...@times-components/depend@2.1.13) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/depend
+
+
+
+
+
 <a name="2.1.12"></a>
 ## [2.1.12](https://github.com/newsuk/times-components/compare/@times-components/depend@2.1.11...@times-components/depend@2.1.12) (2018-11-21)