+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.45](https://github.com/newsuk/times-components/compare/@times-components/pages@1.8.44...@times-components/pages@1.8.45) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [1.8.44](https://github.com/newsuk/times-components/compare/@times-components/pages@1.8.43...@times-components/pages@1.8.44) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [1.8.43](https://github.com/newsuk/times-components/compare/@times-components/pages@1.8.42...@times-components/pages@1.8.43) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [1.8.42](https://github.com/newsuk/times-components/compare/@times-components/pages@1.8.41...@times-components/pages@1.8.42) (2018-12-12)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [1.8.41](https://github.com/newsuk/times-components/compare/@times-components/pages@1.8.40...@times-components/pages@1.8.41) (2018-12-11)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [1.8.40](https://github.com/newsuk/times-components/compare/@times-components/pages@1.8.39...@times-components/pages@1.8.40) (2018-12-11)
 
 **Note:** Version bump only for package @times-components/pages